refactor wip: move renderExplorerPage to baker
